Pull VizPixels up.
